pyproject.toml
[tool.poetry]
name = "redisbench-admin"
version = "0.1.59"
description = "Redis benchmark run helper. A wrapper around Redis and Redis Modules benchmark tools ( ftsb_redisearch, memtier_benchmark, redis-benchmark, aibench, etc... )."
authors = ["filipecosta90 <filipecosta.90@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
redisbench-admin = "redisbench_admin.cli:main"

[tool.poetry.dependencies]
python = "^3.6.1"
humanize = "^2.4.0"
requests = "^2.23.0"
py_cpuinfo = "^5.0.0"
redis = "^3.5.3"
boto3 = "^1.13.24"
tqdm = "^4.46.1"
toml = "^0.10.1"
redistimeseries = "^1.4.3"
redis-py-cluster = "^2.1.0"
paramiko = "^2.7.2"
jsonpath_ng = "^1.5.2"
pysftp = "^0.2.9"
python_terraform = "^0.10.1"
GitPython = "^3.1.12"
PyYAML = "^5.4.0"
wget = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pytest-cov = "^2.9.0"
codecov = "^2.1.4"
black = "^20.8b1"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
